test_function_2/requirements.txt
# Azure Functions for Python
azure-functions>=1.18.0
azure-cosmos>=4.5.0

# Event Hub dependencies
azure-eventhub>=5.11.0

# Optional: for better JSON handling and HTTP requests
requests>=2.31.0